修改公告主键ID为Integer,解决DTO显示时间问题

+++ src/main/java/com/oa/message/dao/INoticeDao.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Component;
 import com.oa.message.entity.Notice;
 
 @Component
-public interface INoticeDao extends PagingAndSortingRepository<Notice, Long> {
+public interface INoticeDao extends PagingAndSortingRepository<Notice, Integer> {
 
 }
